+++ core/src/main/java/hudson/cli/GroovyCommand.java
@@ -25,6 +25,7 @@ package hudson.cli;
 
 import groovy.lang.GroovyShell;
 import groovy.lang.Binding;
+import hudson.cli.util.ScriptLoader;
 import hudson.model.AbstractProject;
 import jenkins.model.Jenkins;
 import hudson.model.Item;
@@ -103,26 +104,7 @@ public class GroovyCommand extends CLICommand implements Serializable {
         if (script.equals("="))
             return IOUtils.toString(stdin);
 
-        return checkChannel().call(new Callable<String,IOException>() {
-            public String call() throws IOException {
-                File f = new File(script);
-                if(f.exists())
-                    return FileUtils.readFileToString(f);
-
-                URL url;
-                try {
-                    url = new URL(script);
-                } catch (MalformedURLException e) {
-                    throw new AbortException("Unable to find a script "+script);
-                }
-                InputStream s = url.openStream();
-                try {
-                    return IOUtils.toString(s);
-                } finally {
-                    s.close();
-                }
-            }
-        });
+        return checkChannel().call(new ScriptLoader(script));
     }
 }
